move make_histogram to the common area
